Maintain deterministic order of CLUSTER SHARDS response
Maintain deterministic order of CLUSTER SHARDS response. Currently we
don't maintain the shards/masters in sorted fashion and hence we get the
order of CLUSTER SHARDS response non-deterministic on different nodes.
Maintain the sorted Masters list of pointers, similar to replicas,
and get rid of <shards, list<nodes>> dict which is not suitable
for sorting. Add TOPOLOGY argument to get the deterministic response
which would remove the replication offset and node health status from
cluster shards response. Sort the nodes based on the node Id. Use it
in proc `cluster_config_consistent` for the test coverage and sanity
purpose.

Signed-off-by: Ram Prasad Voleti <ramvolet@amazon.com>